+++ geotuileur/plugin_main.py
@@ -20,6 +20,9 @@ from geotuileur.__about__ import DIR_PLUGIN_ROOT, __title__, __uri_homepage__
 from geotuileur.gui.dashboard.dlg_dashboard import DashboardDialog
 from geotuileur.gui.dlg_authentication import AuthenticationDialog
 from geotuileur.gui.dlg_settings import PlgOptionsFactory
+from geotuileur.gui.publication_creation.wzd_publication_creation import (
+    PublicationFormCreation,
+)
 from geotuileur.gui.tile_creation.wzd_tile_creation import TileCreationWizard
 from geotuileur.gui.upload_creation.wzd_upload_creation import UploadCreationWizard
 from geotuileur.gui.user.dlg_user import UserDialog
@@ -52,20 +55,22 @@ class GeotuileurPlugin:
         self.action_settings = None
 
         self.toolbar = None
-        self.dlg_authentication = None
         self.dlg_dashboard = None
 
         self.action_authentication = None
         self.action_dashboard = None
         self.action_import = None
         self.action_tile_create = None
+        self.action_publication = None
 
         self.btn_autentification = None
         self.btn_import = None
         self.btn_configuration = None
+        self.btn_publication = None
 
         self.import_wizard = None
         self.tile_creation_wizard = None
+        self.publication_wizard = None
 
     def initGui(self):
         """Set up plugin UI elements."""
@@ -116,6 +121,18 @@ class GeotuileurPlugin:
         )
         self.action_tile_create.triggered.connect(self.tile_creation)
 
+        # Publication
+        self.action_publication = QAction(
+            QIcon(
+                str(
+                    DIR_PLUGIN_ROOT / "resources" / "images" / "icons" / "Publie@2x.png"
+                )
+            ),
+            self.tr("Publication"),
+            self.iface.mainWindow(),
+        )
+        self.action_publication.triggered.connect(self.publication)
+
         # Help
         self.action_help = QAction(
             QIcon(":/images/themes/default/mActionHelpContents.svg"),
@@ -151,8 +168,7 @@ class GeotuileurPlugin:
         self.toolbar.addAction(self.action_dashboard)
         self.toolbar.addAction(self.action_import)
         self.toolbar.addAction(self.action_tile_create)
-
-        # -- Check actions
+        self.toolbar.addAction(self.action_publication)
         self._update_actions_availability()
 
         # -- Processings
@@ -192,16 +208,17 @@ class GeotuileurPlugin:
             self.tile_creation_wizard.finished.connect(self._del_tile_creation_wizard)
         self.tile_creation_wizard.show()
 
-    def _del_tile_creation_wizard(self) -> None:
+    def publication(self):
         """
-        Delete tile creation wizard
+        Open tile creation Wizard
 
         """
-        if self.tile_creation_wizard is not None:
-            self.tile_creation_wizard.deleteLater()
-            self.tile_creation_wizard = None
+        if self.publication_wizard is None:
+            self.publication_wizard = PublicationFormCreation(self.iface.mainWindow())
+            self.publication_wizard.finished.connect(self._del_tile_publication_wizard)
+        self.publication_wizard.show()
 
-    def import_data(self) -> None:
+    def import_data(self):
         """
         Open import data Wizard
 
@@ -220,20 +237,34 @@ class GeotuileurPlugin:
             self.import_wizard.deleteLater()
             self.import_wizard = None
 
+    def _del_tile_creation_wizard(self) -> None:
+        """
+        Delete tile creation wizard
+
+        """
+        if self.tile_creation_wizard is not None:
+            self.tile_creation_wizard.deleteLater()
+            self.tile_creation_wizard = None
+
+    def _del_tile_publication_wizard(self) -> None:
+        """
+        Delete tile publication wizard
+
+        """
+        if self.publication_wizard is not None:
+            self.publication_wizard.deleteLater()
+            self.publication_wizard = None
+
     def authentication(self) -> None:
-        """
-        Open authentification dialog
-
-        """
-
+        """Open authentication dialog."""
         plg_settings = self.plg_settings.get_plg_settings()
 
         if len(plg_settings.qgis_auth_id) == 0:
             dlg_authentication = AuthenticationDialog(self.iface.mainWindow())
-            dlg_authentication.exec()
+            dlg_authentication.exec_()
         else:
             dlg_user = UserDialog(self.iface.mainWindow())
-            dlg_user.exec()
+            dlg_user.exec_()
         self._update_actions_availability()
 
     def _update_actions_availability(self) -> None:
@@ -246,6 +277,7 @@ class GeotuileurPlugin:
 
         self.action_import.setEnabled(enabled)
         self.action_tile_create.setEnabled(enabled)
+        self.action_publication.setEnabled(enabled)
 
     def display_dashboard(self) -> None:
         """